+++ changelog.md
@@ -5,8 +5,11 @@
                         
 [//]: #@corifeus-header:end
 
-### v1.1.12
+### v1.1.13
 * in progress
+
+### v1.1.12
+* allow CodeMirror show toggle height 600px vs all
 
 ### v1.1.11
 * Make to use the tag/branch button like GitHub
@@ -120,7 +123,7 @@ require('codemirror/mode/yaml/yaml');
 
 ---
 
-[**GITLIST**](https://pages.corifeus.com/gitlist) Build v1.1.11 
+[**GITLIST**](https://pages.corifeus.com/gitlist) Build v1.1.12 
 
 [![Like Corifeus @ Facebook](https://img.shields.io/badge/LIKE-Corifeus-3b5998.svg)](https://www.facebook.com/corifeus.software) [![Donate for Corifeus / P3X](https://img.shields.io/badge/Donate-Corifeus-003087.svg)](https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=QZVM4V6HVZJW6)  [![Contact Corifeus / P3X](https://img.shields.io/badge/Contact-P3X-ff9900.svg)](https://www.patrikx3.com/en/front/contact)